+++ src/main/java/com/github/castorm/kafka/connect/http/auth/spi/HttpAuthenticator.java
@@ -1,0 +1,36 @@
+package com.github.castorm.kafka.connect.http.auth.spi;
+
+/*-
+ * #%L
+ * kafka-connect-http
+ * %%
+ * Copyright (C) 2020 CastorM
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+
+import org.apache.kafka.common.Configurable;
+
+import java.util.Map;
+import java.util.Optional;
+
+@FunctionalInterface
+public interface HttpAuthenticator extends Configurable {
+
+    Optional<String> getAuthorizationHeader();
+
+    default void configure(Map<String, ?> map) {
+        // Do nothing
+    }
+}